+++ cargo-wharf-frontend/src/plan.rs
@@ -8,7 +8,7 @@ use serde::{Deserialize, Serialize};
 use buildkit_frontend::Bridge;
 use buildkit_llb::prelude::*;
 
-use crate::config::Config;
+use crate::config::{BaseImageConfig, Config};
 use crate::query::Profile;
 use crate::shared::{tools, CONTEXT, CONTEXT_PATH};
 
@@ -50,7 +50,7 @@ impl RawBuildPlan {
         bridge: &'a mut Bridge,
         config: &'b Config,
     ) -> Result<Self, Error> {
-        let builder = config.builder_image();
+        let builder = config.builder();
 
         let mut args = vec![
             String::from("--manifest-path"),
@@ -65,7 +65,7 @@ impl RawBuildPlan {
                 .into(),
         ];
 
-        if let Some(target) = config.builder_image().target() {
+        if let Some(target) = builder.target() {
             args.push("--target".into());
             args.push(target.into());
         }