+++ cargo-wharf-frontend/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Advanced output image metadata (Dockerfile's `VOLUME`, `EXPOSE`, `LABEL`, `STOPSIGNAL`).
+- Custom image setup commands (Dockerfile's `RUN`).
 
 ### Changed
 - README and usage guide.